+++ src/_interfaces/formatter_widget.js
@@ -3,21 +3,12 @@
     IFace.add('formatter_widget', {
         init: function (options, dtable) {
             this.dtable = dtable;
+            this.options = $.extend(true, {}, this.getDefaults(), options);
         },
-        defOptions: function () {
+        getDefaults: function () {
             return {};
         },
-        getOptions: function (options) {
-
-            var result = this.defOptions();
-
-            if (options != undefined) {
-                result = $.extend(true, {}, this.defOptions(), options);
-            }
-
-            return result;
-        },
-        format: function (value, options) {
+        format: function (columnId, value, values) {
             return value;
         }
     });