+++ src/formatter/widgets/number.widget.js
@@ -1,21 +1,19 @@
 (function (DTableModule, $) {
 
     DTableModule.newModule(DTableModule.MODULE_FORMATTER_WIDGET, "number", {
-        defOptions: function(){
+        getDefaults: function(){
             return {
                 number_format: '0,0.0',
                 language: 'en',
                 force_number: false
             };
         },
-        format: function (value, options) {
-            options = this.getOptions(options);
+        format: function (columnId, value, values) {
 
-            //console.log(parseInt(value));
-            if (!isNaN(parseInt(value)) || options.force_number)
+            if (!isNaN(parseInt(value)) || this.options.force_number)
             {
-                numeral.language(options.language);
-                value = numeral(value).format(options.number_format);
+                numeral.language(this.options.language);
+                value = numeral(value).format(this.options.number_format);
             }
 
             return value;