+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 - adding formatter widgets, merge formatter options with column definition formatter options
 - adding advancved formatter (formatter widget per column)
 - refactoring nunjucks template loader and template interface, now we can add custom templates
+- templates start loading after definition (for partial widget we need definition settings)
+- partial formatter widget
+- advanced formatter with string widget is now default (escape is on)